[CONJS-153] binary protocol support
new methods connection.prepare, connection.execute, connection.unprepare like mysql2 API.

Implementation differ compared to mysql2 :
* support binary batching
* support streaming parameters
* cache is not infinite: using a LRU cache (option `cacheLen`, default to 256, 0 meaning no cache).
* Implement MariaDB 10.6 new feature permitting to skip metadata if cached, server sending only the raw data for faster results

+++ test/integration/test-change-user.js
@@ -6,52 +6,30 @@ const ServerStatus = require('../../lib/const/server-status');
 const Conf = require('../conf');
 
 describe('change user', () => {
-  before((done) => {
-    Promise.all([
-      shareConn.query("DROP USER IF EXISTS ChangeUser@'%'"),
-      shareConn.query("DROP USER IF EXISTS ChangeUser2@'%'")
-    ])
-      .then(() => {
-        return shareConn.query('CREATE DATABASE IF NOT EXISTS test');
-      })
-      .then(() => {
-        return shareConn.query("CREATE USER ChangeUser@'%' IDENTIFIED BY 'm1P4ssw0@rd'");
-      })
-      .then(() => {
-        return shareConn.query(
-          'GRANT SELECT,EXECUTE ON `' + Conf.baseConfig.database + "`.* TO ChangeUser@'%'"
-        );
-      })
-      .then(() => {
-        return shareConn.query("CREATE USER ChangeUser2@'%' IDENTIFIED BY 'm1SecondP@rd'");
-      })
-      .then(() => {
-        return shareConn.query(
-          'GRANT SELECT,EXECUTE ON `' +
-            Conf.baseConfig.database +
-            "`.* TO ChangeUser2@'%' with grant option"
-        );
-      })
-      .then(() => {
-        return shareConn.query('FLUSH PRIVILEGES');
-      })
-      .then(() => done())
-      .catch((err) => done());
+  before(async () => {
+    if (process.env.srv !== 'maxscale' && process.env.srv !== 'skysql-ha') {
+      await shareConn.query("DROP USER IF EXISTS ChangeUser@'%'");
+      await shareConn.query("DROP USER IF EXISTS ChangeUser2@'%'");
+      await shareConn.query('CREATE DATABASE IF NOT EXISTS test');
+      await shareConn.query("CREATE USER ChangeUser@'%' IDENTIFIED BY 'm1P4ssw0@rd'");
+      await shareConn.query(
+        'GRANT SELECT,EXECUTE ON `' + Conf.baseConfig.database + "`.* TO ChangeUser@'%'"
+      );
+      await shareConn.query("CREATE USER ChangeUser2@'%' IDENTIFIED BY 'm1SecondP@rd'");
+      await shareConn.query(
+        'GRANT SELECT,EXECUTE ON `' +
+          Conf.baseConfig.database +
+          "`.* TO ChangeUser2@'%' with grant option"
+      );
+      await shareConn.query('FLUSH PRIVILEGES');
+    }
   });
 
-  after((done) => {
-    shareConn
-      .query("DROP USER IF EXISTS ChangeUser@'%'")
-      .then(() => {
-        return shareConn.query("DROP USER IF EXISTS ChangeUser2@'%'");
-      })
-      .then(() => {
-        return shareConn.query('FLUSH PRIVILEGES');
-      })
-      .then(() => {
-        done();
-      })
-      .catch((err) => done());
+  after(async () => {
+    if (process.env.srv !== 'maxscale' && process.env.srv !== 'skysql-ha') {
+      await shareConn.query("DROP USER IF EXISTS ChangeUser@'%'");
+      await shareConn.query("DROP USER IF EXISTS ChangeUser2@'%'");
+    }
   });
 
   it('basic change user using callback', function (done) {
@@ -81,7 +59,7 @@ describe('change user', () => {
   });
 
   it('wrong charset', function (done) {
-    if (process.env.MAXSCALE_TEST_DISABLE) this.skip();
+    if (process.env.srv == 'maxscale' || process.env.srv === 'skysql-ha') this.skip();
     if (!shareConn.info.isMariaDB()) this.skip();
     base.createConnection().then((conn) => {
       conn
@@ -323,7 +301,6 @@ describe('change user', () => {
               shareConn.info.isMariaDB() &&
               shareConn.info.hasMinVersion(10, 2, 2) &&
               process.env.srv !== 'maxscale' &&
-              process.env.srv !== 'skysql' &&
               process.env.srv !== 'skysql-ha'
             ) {
               assert.equal(conn.info.database, 'test');
